requirements.txt
aiml

pip install django-user-tracking
tracking
django-crum
